feat: many changes to the project
Fully implemented the values:list service with db implementation towards
postgres database. Both service and dao modules were implemented fully
using the sqlx library.

The service is now able to handle requests for listing values, and the
dao is able to handle the database interactions. The k6 test script
was updated to test the new service endpoint.

Added a lot of documentation, but some places still need more. In
particular the from codes should better described.

Logging is not yet implemented, but the structure is in place for future
work.

Prometheus metrics are implemented, but is still missing memory and
database metrics.

Breaking changes: None

Resolves: None