feat: Adds logging
Adds log4rs to the project for better logging capabilities. An example
configuration file is provided in `test_config/logfile.yaml`. The
logging configuration can be customized as needed.

Tracing logging is also added to the project, which allows for more
detailed and structured logging. The `tracing` crate is used for this
purpose.

Adds a few model changes to allow the models to be logged.

Adds new parameter to the database config for logging slow connection
aquires.

Breaking changes: None

Resolves: None