feat: Add httpsignatures support
Adds support for HTTP Signatures as per RFC 9421, allowing for secure
and verifiable API requests. Currently ed25519, ecdsa and rsa signatures
are supported. Hmac and json web token signatures are not supported yet.

Supported functionality.
- Required headers
- Required headers if body is included.
- Expires required
- Expires verification
- Created required

Currently only request-target and method are supported of the derived
signature fields. The rest of the signature fields are not yet
supported, but should be fairly easy to add later.

It uses the ring library to verify the signatures.

This is just an initial commit as much work remains to be done.

Future work:
- Add support for more signature fields.
- Add support for HMAC and JWT signatures.
- Add better support for different required headers.
- Add independent error handling for signature verification.
- Add tests for signatures generated by different sources.
- Add support for signature generation.

Breaking changes: None

Resolves: None