feat: Httpsignature fixes and tests
This commit includes several fixes and improvements to the HTTPSignature
implementation, along with the addition of comprehensive tests to
ensure.

Adds keys for rsa, ecdsa, and ed25519 algorithms, enhancing the
test coverage for the different signature algorithms. The tests algorithms
includes a hmac-256 algorithm test, which was previously missing.

The signature verification process has been added to all endpoints,
ensuring that all API requests are properly authenticated.

The application now uses openssl for signature verification as the ring
library is using BoringSSL and says it's not suitable for production
use. Ring is still used for generating signatures in test as it allows
for verifying signatures generated using another library.

The configuration has been updated to include allowd keys, currently
only pem is supported.

The config now allows for setting requirements for the signature.

Breaking changes: None

Resolves: None