feat: Adds httpsignatures support on responses
Adds support for HTTP signatures on responses, allowing clients to
verify the integrity and authenticity of the response data. This
feature enhances security by ensuring that the response has not been
tampered with during transmission.

This adds a derived value @status to the signature verification which
can only be used for responses. The others have been changed to
optional to allow for more flexibility in the signature verification.

Also fixes a minor typo in test data.

Breaking changes: None

Resolves: None