feat: Implements algorithm for signature generation
Implements functionality to generate signatures using the algorithms
specified in rfc-9421. The implementation does not include generating
the signature base, this will be included later. Currently it uses
openssl to generate the signatures. Performance testing of the code
has not been done yet.

Algorithm has been added to the keys for signature verification
this allows verification of incorrect algorithms to fail faster
improving performance.

Changes logging level for errors that should not occur to warning
level rather than debug.

Breaking changes: Requires adding algorithm to keys in the config.

Resolves: None