refactor: Add signature generation
This commit introduces the generation of HTTP signatures for all
relevant endpoints and conditions when returning responses. The
generation supports both RSA, ECDSA, ED25519 and HMAC signatures.

Currently the keyid is hardcoded, this will be made configurable in
a future commit.

This commit also includes various refactoring changes across multiple files
to enhance code structure and readability. This also means adding both
digest and signature verification to all relevant endpoints.

Improves code readability on the endpoints so that it separates the
business part from the technical part.

There are some duplication of code after moving digest verification out
of middleware. This should be addressed in future refactoring by moving
digest and signature verification back to middleware.

Changes the endpoints to use payload instead of serde objects
directly. This is neccesary since we need the raw body for digest
verification.

Breaking changes: None

Resolves: None