feat: Bug fixes and code cleanup
This commit focuses on code cleanup, optimization and fixing a few
minor bugs. The changes include:

1. Logging improvements have been changed where necessary and added so
that the application logs more useful information. In particular logging
what the signature error is.

2. Testing found a bug where the generated output signature was
generated and added twice. The cause for this is bad code design. This
issue fixes this by cleaning up the code by ensuring that the generation
is only done at the endpoints.

3. Code cleanup and optimization has been done at the endpoints to fix
a few minor bugs where the incorrect response was returned due to
incorrect error handling. The error caused successes to be reported
when they should not have been.

4. Fixes a bug when the created and expired was optional in the request,
this caused all requests to fail. This has been fixed by not adding
them to the signature params.

5. Changes the key configuration so it's a vector instead of a map which
is then converted to a map inside the application. This improves the
readability of the configuration files,

Breaking changes: The configuration file format has changed for the
keys. Keyid is now required and the keys are now a list instead of a
map.

Resolves: None