fix: Fix signature params
Fixes an issue where signature base used incorrect signature params. The
cause for this was the assumption that the parameters were always in the
same order, but this is correct as the order is not guaranteed. The fix
is therefore to not generate the signature params, but to use the same
as in the input.

Breaking changes: None as this only an internal change in the signature
verification.

Resolves: None