feat: Add Jaeger tracing support and Docker setup
Adds opentelemetry and jaeger support for distributed tracing. The code
removes the previous loggind currently only uses jaeger. A later
improvement will send metrics to prometheus. The features uses the
official opentelemetry crates for the and tracing_opentelemetry crate
for connecting the tracing crate to opentelemetry.

Current code needs some cleanup and refactoring, but the main
functionality is in place.

A jaeger docker setup has been added to the docker folder. The
postgres setup has been added to the folder.

The logging configuration has been updated to only opentelemetry. The
jaeger url is configurable in the config file.

Breaking changes: None

Resolves: None