feat: telemetry improvements
Improves telemetry by transmitting metrics to Prometheus. A
configuration prometheusUrl has been added to the config file. If added
the application will push metrics to the Prometheus application. If not
added this will be ignored.

Added prometheus to the telemetry docker. It will now run both
jaeger and prometheus.

Changes the trace metrics instruments to log only on debug level as this
was too verbose for info level.

Adds trace to both acquire db connection and begin transaction.

Future improvements could be to send logs to a log management system.
The logs could then be searched and monitored.

Breaking changes: None

Resolves: None