feat: Updates logging
Adds support for Loki for storing and querying logs. A url has been
added in the configuration for Loki. Docker configuration files have
been added for Loki and Promtail.

Adds prometheus configuration for testing.

Updates tracing so that begin transaction, commit transaction, and
rollback are traced for time in jaeger.

Changes the url for jaeger and prometheus yto no longer be optional.

Breaking change: The configuration file must be updated to include
the new fields for Loki and the updated fields for Jaeger and
Prometheus if missing.

Future improvements: Add back directives for logging.

Resolves: None